feat: Add useProjects hook for fetching and managing projects
This commit adds the `useProjects` hook to the `useProjects.ts` file. The hook is responsible for fetching projects using the RTK Query `useGetProjectsQuery` hook and saving them in the Redux state using the `setProjects` action from the `projectsSlice`. The fetched projects can be accessed from the Redux state using the `projects` selector.

Note: The commit message has been modified to adhere to the established commit message conventions.