feat: Update page titles and header styles in FormScreen and ProjectDetails components
This commit updates the page titles and header styles in the FormScreen and ProjectDetails components. In the FormScreen component, the title is changed to display the form name instead of the form ID. In the ProjectDetails component, the title is updated to show the project name. Additionally, the headerShadowVisible property is set to false in the ProjectDetails component to remove the shadow from the header.

Note: The commit message has been modified to adhere to the established commit message conventions.